+++ Controller/SuggestionProviderController.php
@@ -14,6 +14,7 @@ namespace Symfony\Cmf\Bundle\SeoBundle\Controller;
 use Symfony\Bundle\TwigBundle\Controller\ExceptionController;
 use Symfony\Cmf\Bundle\SeoBundle\SuggestionProviderInterface;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestMatcherInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\FlattenException;
 use Symfony\Component\HttpKernel\Log\DebugLoggerInterface;
@@ -34,6 +35,45 @@ class SuggestionProviderController extends ExceptionController
     protected $suggestionProviders = array();
 
     /**
+     * Contains the list of templates defined in the error section of the
+     * cmf_seo configuration.
+     *
+     * Looks like
+     *
+     *  array(
+     *      'format' => 'MyBundle:Error:exception.format.twig,
+     *  )
+     *
+     * @var array
+     */
+    private $templates;
+
+    /**
+     * @var RequestMatcherInterface
+     */
+    private $exclusionRequestMatcher;
+
+    /**
+     * @param \Twig_Environment       $twig
+     * @param bool                    $debug
+     * @param RequestMatcherInterface $requestMatcher The exclusion matcher to decider whether a route should be handled
+     *                                                by this error handling. It uses the defined exclusion_rules in the
+     *                                                error configuration.
+     * @param array                   $templates  Containing the configured templates to use in custom error cases.
+     */
+    public function __construct(
+        \Twig_Environment $twig,
+        $debug,
+        RequestMatcherInterface $requestMatcher,
+        $templates
+    ) {
+        $this->templates = $templates;
+        $this->exclusionRequestMatcher = $requestMatcher;
+
+        parent::__construct($twig, $debug);
+    }
+
+    /**
      * @param Request              $request
      * @param FlattenException     $exception
      * @param DebugLoggerInterface $logger
@@ -47,8 +87,12 @@ class SuggestionProviderController extends ExceptionController
         $_format = 'html'
     ) {
         $code = $exception->getStatusCode();
+        if (404 !== $code || $this->exclusionRequestMatcher->matches($request)) {
+            return parent::showAction($request, $exception, $logger, $_format);
+        }
 
-        if (404 !== $code) {
+        $templateForSuggestion = $this->getTemplateForSuggestions($_format);
+        if (null === $templateForSuggestion) {
             return parent::showAction($request, $exception, $logger, $_format);
         }
 
@@ -64,10 +108,11 @@ class SuggestionProviderController extends ExceptionController
 
         return new Response(
             $this->twig->render(
-                $this->findTemplate($request, $_format, $code, $this->debug),
+                $templateForSuggestion,
                 array(
                     'status_code'    => $code,
                     'status_text'    => isset(Response::$statusTexts[$code]) ? Response::$statusTexts[$code] : '',
+                    'message'        => $exception->getMessage(),
                     'exception'      => $exception,
                     'logger'         => $logger,
                     'currentContent' => $currentContent,
@@ -85,5 +130,21 @@ class SuggestionProviderController extends ExceptionController
     public function addSuggestionProvider(SuggestionProviderInterface $matcher, $group)
     {
         $this->suggestionProviders[] = array('provider' => $matcher, 'group' => $group);
+    }
+
+    /**
+     * You can define your templates for each format in the bundle's configuration.
+     *
+     * @param string $format
+     *
+     * @return string
+     */
+    private function getTemplateForSuggestions($format = 'html')
+    {
+        if (!isset($this->templates[$format])) {
+            return null;
+        }
+
+        return $this->templates[$format];
     }
 }